Split the evdev backend into modules along its own seams
evdev.rs had grown to 1825 lines covering five distinct jobs. It is now a
directory, split where the code already divided:

  mod.rs      117  the module doc, the shared EV_KEY taxonomy, re-exports
  router.rs   563  the pump and the reconcile loop that arms it
  frame.rs    638  poll set, drain, routing, frame composition (372 tests)
  cursor.rs   256  the authoritative cursor (107 tests)
  devices.rs  233  enumeration, scanner thread, dead-node purge (53 tests)
  probe.rs    168  compositor cursor position, desktop bounds
  uinput.rs    85  the two virtual devices

Which EV_KEY codes are buttons and which are keyboard usages is the one
piece of knowledge every module needs — frame routes on it, uinput
declares its complement, cursor and devices index the held mask with
it — so the three ranges stay in the parent module. Everything else has
one owner.

The public surface is unchanged: available, run, kwin_cursor_pos and the
three types benches/evdev_pump drives are re-exported from mod.rs, and
no file outside the directory needed touching.

Behaviour is unchanged too. Statement by statement the bodies are the
same; what moved is visibility markers, imports, and three seams that
only exist because the code is now in separate files:

- Mouse and Node type aliases for the (path, device[, accel]) tuples
  the signatures used to spell out;
- the poll set and the dead-slot list stay private to frame, behind
  slots/poll/revents/mark_dead/dead_slots/clear_dead — which also puts
  the unsafe poll(2) call next to the buffer that owns the fds;
- the scanner thread and the arm-time enumeration both go through
  devices::enumerate/spawn_scanner rather than being inlined in arm.

The reusable buffers and their numbers are intact: benches/evdev_pump
still reports 0 allocations per report and 8.5 ns for the ordinary
motion cycle. 99 tests pass, moved to the module each covers. cargo fmt
--check clean, clippy unchanged from the baseline (same 15 library
warnings, on the same pre-existing lines).
